+++ doc/conf.py
@@ -74,7 +74,7 @@ copyright = '2014-2016, Louis Paternault'
 # built documents.
 #
 # The short X.Y version.
-version = "0.4.0"
+version = "0.4.1"
 # The full version, including alpha/beta/rc tags.
 release = version